present: Fixed a bug that style was broken when slide printing.
slide.js in the present package automatically scale up and scale down
the slide according to the size of the browser.
It is using "transoform property" of stylesheet.

However, at the time of printing, the "transform property" remains, and
the stylesheet for printing is broken.
I modified it so that "transform property" is reset when printing.

Fixes https://github.com/golang/go/issues/29480