go/packages: stop parsing files if the context is cancelled
`stamblerre/gocode`, which is the fork of `mdempsky/gocode` that
supports Go 1.11 modules, uses `packages.Load` to load all of the
Go code for, among other things, IDE autocomplete support.  A
common aspect to IDE autocomplete is asking for suggestions very
frequently (perhaps at every character typed).  Once the user
has typed another character, the previous request for autocomplete
is invalid and can be cancelled.

`packages.Load` already stops its `go list` component if the
context is cancelled, but if the context is cancelled afterward,
then it will parse all of the files that it found.  This change
stops the parsing of Go files once it detects that the context
has been cancelled.

This change dramatically improves the performance of the
`stamblerre/gocode` fork when requests have been cancelled.