internal/lsp: get the correct symbol kind for the type alias and type definition
For the original implementation, there are just two
symbol kinds, `StructSymbol` and `InterfaceSymbol`.
And it won't take the essential type into account. Like
the code bellow, the symbol kind of `alias` will be
`StructSymbol`, that's not correct.

```
type Node interface{
}

type alias = Node
```

This patch consider the essential type to give a more
precise result of the symbol kind.