[elastic] Get the correct symbol kind for the type alias and type definition (#6)
For the original implementation, there is no consideration about the
type alias, so for the type alias or the type definition, there won't
valid symbol kind. This patch strip off the alias layer to get the
essential type. And get the correct symbol kind based on that.

Btw, it have to be noted that there are chances that server will return
invalid symbol kind to the code plugin. This is not a good practice.