[elastic] Manage the deps. (#7)
This is the initial commit of the deps management. Considering the
consistency and the implementation, use the '$GOPATH/pkg/mod', which is
the default location to hold the module deps, to store the deps.

Use a specified folder under '.../kibana/data/code' is difficult to
implement, we have to adjust the package resolve mechanism
correspondingly. Another option, i.e. creating the 'vendor' folder in
every repo has several drawbacks as bellow:
 - Storage waste
 - Append every go command, mainly `go list`, with `-mod=vendor`. And we
   need to expand initialize options to tell the lsp to treat current
   repo with `-mod=vendor`.

For now, we can handle follow scenarios:
 - Repos with neither `go.mod` nor `vendor`.
 - Repos with dependence control files, like `vendor/vendor.json`,
 `Gopkg.lock`.
 - Repos with `go.mod`, actually there is nothing we should do about
   this scenario.

We can't handle follow scenarios:
 - Repos just have `vendor` folder.

Btw, there are performance issues, `go list` takes a long time to
download the deps.