internal/lsp: add definition support for import specs
Now we consider the "definition" of an *ast.ImportSpec to be the
package being imported. We have to pick one of the package's files, so
we choose the longest one, hoping it is the most interesting.

I also added a nil check for the packge in SignatureHelp. This panics
for me occasionally.