internal/lsp: don't complete test objects in non-test files
go/packages already treats code and code+test as two separate
packages, but we were caching packages by only their package path,
which is not unique. If you loaded a test file first, all the non-test
files in the package would become associated to the code+test package,
so the non-test files would behave as if they could see objects
defined in the test files.

Now we cache packages by their path and whether they contain tests or
not. We also avoid associating the code+test package to the non-test
files so that opening a non-test file will properly load the non-test
package instead of reusing the code+test package.

This means we will now parse and type check non-test files twice,
which will impact memory usage and perhaps responsiveness in some
scenarios.