internal/lsp: fix error handling after a file is deleted
In a previous commit I made error checking stricter when parsing a
package's files. However, when a file is deleted from a package, our
parsing code starts to get "no such file" errors, which if treated as
a fatal error prevents the package from getting parsed. It doesn't
seem like there is a good way to detect when a file is deleted and
refresh the metadata, so revert the stricter error handling for now.