[elastic] Sync to the upstream (#19)
The upstream trims the AST of the dependencies to reduce the memory
usage. That means for the symbol defined in dependencies langserver will
reparse the source files to get the full ASTs, this approach will
introduce overhead for the cross-repo jump.

Upstream has supported the workspace folder change, the `ManageDeps`
should apply to the workspace folder change API to handle the repository
that owns `vendor` file or contains multiple modules.