[elastic] Add go ci. (#22)
For now, the docker file and the script just for the unit test. Once the
go language server ready to be released, we should add the package script
and enrich the docker file.

The fact I want to emphasize here is that go langserver rely on
unfinished gopls, which located in the `internal` folder. That's why the
go langserver based on the whole `tools` repo. Once the gopls finished,
we should refactor the functionality we added.

Btw, go packagetest has issues when we import non-builtin packages, I
remove these packages temporarily to make the test sucesses.