[elastic] Automatically deploy the artifacts
We created the tags manually in initial design, it is not the
perfect approach. For now, we create tags for every PR-merge
to trigger the artifacts generate.

For the commits on master branch, the tag will be the short
version of SHA1. For the commits on non-master branch, the
tags of the other branches will be 'brach-name'-'short-SHA1'.

For the ReleaseManager, it will fetch the artifacts through
the short version SHA1 of the lastest commit.