tools/internal/tool: refactor tool.Main() for testabilty
In it previous implementation, tool.Main was meant to be called from
an applications main().  If it encountered an error it would print the
error to standard error and exit with a non-zero status (2).

It is also called recursively and in various test functions.

Exiting on an error makes testing difficult, unnecessarily.

This change breaks the functionality into to parts: an outer
tool.Main() that is intended to be called by main() functions and
an inner Run() that is used by Main() and by test functions.

None of the existing test functions use Run()'s error value, but the
failure case tests for the command line invocation of rename (#194878)
require it.

Fixes #34291