go/analysis/unitchecker: add erroras analysis to align with go vet
The unitchecker according to its comments is an example of replicating
the analysis performed by go vet using the unitchecker. The example is
missing the erroras analysis that was added to go vet in
golang/go@9f76566. This change brings the example back inline with the
analysis performed by go vet.